fix(members): restore membership onboarding build
